+++ CHANGELOG.md
@@ -6,6 +6,8 @@ versions.
 
 ## [Unreleased]
 
+## [0.1.0a3] - 2026-07-16
+
 ### Added
 
 - Support Windows in the Python distribution and provide standalone x64 and
@@ -43,6 +45,7 @@ versions.
 - Require explicit confirmation and process creation-time validation before
   terminating a process.
 
-[Unreleased]: https://github.com/LeonPure/portboard/compare/v0.1.0a2...HEAD
+[Unreleased]: https://github.com/LeonPure/portboard/compare/v0.1.0a3...HEAD
+[0.1.0a3]: https://github.com/LeonPure/portboard/compare/v0.1.0a2...v0.1.0a3
 [0.1.0a2]: https://github.com/LeonPure/portboard/compare/v0.1.0a1...v0.1.0a2
 [0.1.0a1]: https://github.com/LeonPure/portboard/releases/tag/v0.1.0a1